# src/main/resources/openapi/parser-api.openapi.yml
openapi: 3.0.0
info:
  title: Data Harvest Parser
  version: 0.0.1
servers:
  - url: http://localhost:8082/api
    description: Development Server

paths:
  /parser/preset:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Parser
      parameters:
        - name: userId
          in: query
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      summary: Get All Parser Settings Presets by User Id
      operationId: getParserSettingsByUserId
      responses:
        '201':
          description: All Users Parser Settings Presets was Retrieved Successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserParserSettingsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: Internal Server Error
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Parser
      parameters:
        - name: userId
          in: query
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      summary: Create Parser Settings Preset
      operationId: createParserSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserParserSettingsOpenApi"
      responses:
        '201':
          description: Parser Settings Preset was Created Successfully
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: Internal Server Error


  /parser/preset/{presetId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Parser
      parameters:
        - name: presetId
          in: path
          description: Preset id
          required: true
          schema:
            type: integer
            format: int64
      summary: Get Parser Settings Preset by Id
      operationId: getParserSettingsById
      responses:
        '201':
          description: Parser Settings was Retrieved Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserParserSettingsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: Internal Server Error
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Parser
      parameters:
        - name: presetId
          in: path
          description: Preset id
          required: true
          schema:
            type: integer
            format: int64
      summary: Delete Parser Settings Preset by Id
      operationId: deleteParserSettingsById
      responses:
        '201':
          description: Parser Settings was Deleted Successfully
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: Internal Server Error
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Parser
      parameters:
        - name: presetId
          in: path
          description: Preset id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParserResultOpenApi"
      summary: Run Parser With Concrete Preset by its Id and Saving New Parsing History With Current Date
      operationId: runParser
      responses:
        '201':
          description: Parser Settings was Deleted Successfully
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: Internal Server Error



  /parser/preset/{presetId}/download:
    get:
      security:
        - bearerAuth: [ ]
      summary: Download Results File by the Concrete Preset ID
      tags:
        - Parser
      parameters:
        - in: path
          name: presetId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Parser Preset ID
      operationId: downloadFile
      responses:
        200:
          description: The file download was successful
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Internal Server Error


  /storage:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - storage
      summary: Retrieving The Whole Storage With Items of User with Concrete Id
      parameters:
        - name: userId
          in: query
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      operationId: getStorageByUserId
      responses:
        200:
          description: The Storage was retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageOpenApi"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Internal Server Error
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - storage
      summary: Update The Storage With Items of User with Concrete Id
      parameters:
        - name: userId
          in: query
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      operationId: updateStorageByUserId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageOpenApi"
      responses:
        200:
          description: The Storage was retrieved successfully
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Internal Server Error


  /storage/{storageId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - storage
      summary: Retrieving The Whole Storage With Items of Storage by its Id
      parameters:
        - name: storageId
          in: path
          description: Storage id
          required: true
          schema:
            type: string
            format: uuid
      operationId: getStorageById
      responses:
        200:
          description: The Storage was updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageOpenApi"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Internal Server Error
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - storage
      summary: Update The Storage With Items By its Id
      parameters:
        - name: storageId
          in: path
          description: Storage id
          required: true
          schema:
            type: string
            format: uuid
      operationId: updateStorageById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageOpenApi"
      responses:
        200:
          description: The Storage was updated successfully
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Internal Server Error

  /storage/{storageId}/folder/{folderId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Storage
      summary: Get Folder by Folder ID
      operationId: getFolderByFolderId
      parameters:
        - name: storageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of Storage to retrieve
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the folder to retrieve
      responses:
        '200':
          description: The Folder was retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderOpenApi'
        401:
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Folder not found
        '500':
          description: Internal Server Error

    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Storage
      summary: Update a folder by folder ID
      operationId: updateFolderById
      parameters:
        - name: storageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of Storage to update
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the folder to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderOpenApi"
      responses:
        '200':
          description: The folder was updated successfully
        401:
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Folder with such id not found
        '500':
          description: Internal Server Error

    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Storage
      summary: Delete a folder by folder ID
      operationId: deleteFolderById
      parameters:
        - name: storageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of Storage to delete
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the folder to delete
      responses:
        '200':
          description: The folder was deleted successfully
        401:
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Folder with such id not found
        '500':
          description: Internal Server Error

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    UnauthorizedError:
      description: Access token is missing or invalid

  schemas:

    UserParserSettingsOpenApi:
      allOf:
        - $ref: '#/components/schemas/StorageItemOpenApi'
        - type: object
          required:
            - storageId
            - firstPageUrl
            - numOfPagesToParse
            - className
            - tagName
            - cssSelectorNextPage
            - header
            - elementLocators
            - parsingHistory
          properties:
            id:
              type: string
              format: uuid
            storageId:
              type: string
              format: uuid
            name:
              type: string
            tags:
              type: array
              items:
                type: string
            firstPageUrl:
              type: string
            numOfPagesToParse:
              type: integer
              format: int32
            className:
              type: string
            tagName:
              type: string
            cssSelectorNextPage:
              type: string
            header:
              type: array
              items:
                type: string
            elementLocators:
              type: array
              items:
                $ref: '#/components/schemas/ElementLocatorOpenApi'
            parsingHistory:
              type: array
              items:
                $ref: '#/components/schemas/ParserResultOpenApi'
            parentFolderId:
              type: string
              format: uuid


    ParserResultOpenApi:
      type: object
      required:
        - date
        - outputFileType
        - userParserSettingsId
        - linkToDownloadResults
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        linkToDownloadResults:
          type: string
        outputFileType:
          type: object
          $ref: "#/components/schemas/OutputFileTypeOpenApi"
        userParserSettingsId:
          type: string
          format: uuid

    ElementLocatorOpenApi:
      type: object
      required:
        - name
        - elementType
        - pathToLocator
        - userParserSettingsId
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        elementType:
          type: object
          $ref: "#/components/schemas/ElementTypeOpenApi"
        pathToLocator:
          type: string
        extraPointer:
          type: string
        userParserSettingsId:
          type: string
          format: uuid

    FolderOpenApi:
      allOf:
        - $ref: '#/components/schemas/StorageItemOpenApi'
        - type: object
          required:
            - storageId
            - parentFolderId
            - folderItems
          properties:
            id:
              type: string
              format: uuid
            storageId:
              type: string
              format: uuid
            parentFolderId:
              type: string
              format: uuid
            folderItems:
              type: array
              items:
                $ref: '#/components/schemas/StorageItemOpenApi'

    StorageOpenApi:
      type: object
      required:
        - id
        - userId
        - storageItems
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: integer
          format: int64
        storageItems:
          type: array
          items:
            $ref: '#/components/schemas/StorageItemOpenApi'

    StorageItemOpenApi:
      type: object
      properties:
        name:
          type: string
        tags:
          type: array
          items:
            type: string

    OutputFileTypeOpenApi:
      type: string
      enum:
        - XLSX
        - CSV

    ElementTypeOpenApi:
      type: string
      enum:
        - XPATH
        - TAG_ATTR
        - CSS